+++ src/components/atoms/RadioButton/index.jsx
@@ -32,10 +32,10 @@ ${props => {
 `
 
 
-const RadioBox = ({ size, name, onClick, options, checkedValue, disabled }) => (
+const RadioBox = ({ justify, size, name, onClick, options, checkedValue, disabled }) => (
     //options = selectBox 목록 DataType = Array
     <>
-        <RadioGroup>
+        <RadioGroup justify={justify}>
 
             {options.map((i, index) =>
                 <RadioLabel size={size} key={index}><Radio key={index} value={i} disabled={disabled.indexOf(i) !== -1 ? true : false} defaultChecked={(checkedValue === i) ? true : false} onClick={onClick} name={name} />{i}</RadioLabel>
