+++ src/containers/redux/pages/profile/ContentContainer.jsx
@@ -4,15 +4,16 @@ import ACTION from "../../../../store/actions/action"
 
 const mapStateToProps = (state) => {
     return {
-        username: state.user_reducer.username,
         role: state.user_reducer.role,
-        logined: state.login_reducer.logined,
+        username: state.user_reducer.username,
+        realname: state.user_reducer.realname,
+        email: state.user_reducer.email,
         emailValidation: state.user_reducer.emailValidation,
         firstRegion: state.user_reducer.firstRegion,
         secondRegion: state.user_reducer.secondRegion,
-        email: state.user_reducer.email,
-        imgUrl: state.user_reducer.imgUrl || null
-
+        imgUrl: state.user_reducer.imgUrl,
+        sex: state.user_reducer.sex,
+        phone: state.user_reducer.phone
     }
 }
