+++ src/pages/Profile/index.jsx
@@ -2,7 +2,7 @@
  * @author : chaeeun
  * @Date : 2021-03-17 15:43:10 
  * @Last Modified by: euncherry
- * @Last Modified time: 2021-03-22 19:46:27
+ * @Last Modified time: 2021-03-22 19:45:53
  */
 import React from "react"
 import ContentContainer from "../../containers/redux/pages/profile/ContentContainer"
